+++ ActiveRecord.php
@@ -231,8 +231,10 @@ abstract class ActiveRecord extends BaseActiveRecord
             }
         }
         $newId = static::getCollection()->insert($values);
-        $this->setAttribute('_id', $newId);
-        $values['_id'] = $newId;
+        if ($newId !== null) {
+            $this->setAttribute('_id', $newId);
+            $values['_id'] = $newId;
+        }
 
         $changedAttributes = array_fill_keys(array_keys($values), null);
         $this->setOldAttributes($values);