+++ file/ActiveRecord.php
@@ -128,8 +128,10 @@ abstract class ActiveRecord extends \yii\mongodb\ActiveRecord
         } else {
             $newId = $collection->insert($values);
         }
-        $this->setAttribute('_id', $newId);
-        $values['_id'] = $newId;
+        if ($newId !== null) {
+            $this->setAttribute('_id', $newId);
+            $values['_id'] = $newId;
+        }
 
         $changedAttributes = array_fill_keys(array_keys($values), null);
         $this->setOldAttributes($values);