+++ tests/file/StreamWrapperTest.php
@@ -80,4 +80,38 @@ class StreamWrapperTest extends TestCase
 
         $this->assertEquals('test content', stream_get_contents($resource));
     }
+    
+    public function testSeek()
+    {
+        $connection = $this->getConnection();
+        $this->mockApplication([
+            'components' => [
+                'mongodb' => $connection
+            ],
+        ]);
+
+        $connection->registerFileStreamWrapper(true);
+        $databaseName = $connection->getDefaultDatabaseName();
+
+        $url = "gridfs://{$databaseName}.fs?filename=test.txt";
+        $resource = fopen($url, 'w');
+        fwrite($resource, 'begin end');
+        fclose($resource);
+        
+        $url = "gridfs://{$databaseName}.fs?filename=test.txt";
+        $resource = fopen($url, 'r');
+        $data = fgets($resource);
+        
+        fseek($resource, 0);
+        $position = ftell($resource);
+        $this->assertEquals(0, $position);
+        
+        fseek($resource, 2, SEEK_CUR);
+        $position = ftell($resource);
+        $this->assertEquals(2, $position);
+        
+        fseek($resource, 0, SEEK_END);
+        $position = ftell($resource);
+        $this->assertEquals(9, $position);
+    }
 }